fix: cli
