+++ pyproject.toml
@@ -1,38 +1,30 @@
-[project]
+[tool.poetry]
 name = "pysatl-experiment"
 version = "0.1.0"
 description = "package for PySATL math statistics experiments"
-authors = [
-    {name = "Ivan Pokhabov", email = "vanek3372@gmail.com"},
-]
-license = {text = "MIT"}
+authors = ["Ivan Pokhabov <vanek3372@gmail.com>"]
+license = "MIT"
 readme = "README.md"
-requires-python = ">=3.10,<3.13"
-dependencies = [
-    "numpy>=1.25.1",
-    "scipy>=1.11.2",
-    "matplotlib>=3.8.0",
-    "tqdm>=4.66.2",
-    "fpdf2>=2.7.8",
-    "jinja2>=3.1.5",
-    "xhtml2pdf>=0.2.17",
-    "pandas>=2.2.1",
-    "sqlalchemy>=2.0.36",
-    "python-rapidjson==1.20",
-    "jsonschema==4.23.0",
-    "rich==13.9.4",
-    "click>=8.2.1",
-    "dacite==1.9.2",
-    "pysatl-criterion @ ./pysatl_criterion"
-]
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry]
 package-mode = true
 packages = [{include = "pysatl_experiment"}]
+
+[tool.poetry.dependencies]
+python = ">=3.10,<3.13"
+numpy = ">=1.25.1"
+scipy = ">=1.11.2"
+matplotlib = ">=3.8.0"
+tqdm = ">=4.66.2"
+fpdf2 = ">=2.7.8"
+jinja2 = ">=3.1.5"
+xhtml2pdf = ">=0.2.17"
+pandas = ">=2.2.1"
+sqlalchemy = ">=2.0.36"
+python-rapidjson = "==1.20"
+jsonschema = "==4.23.0"
+rich = "==13.9.4"
+click = ">=8.2.1"
+dacite = "==1.9.2"
+pysatl-criterion = {path = "./pysatl_criterion"}
 
 [tool.poetry.group.dev.dependencies]
 markdown = "3.8"
@@ -40,7 +32,6 @@ mkdocs = "1.6.1"
 mkdocs-material = "9.6.14"
 mdx-truly-sane-lists = "1.3"
 pymdown-extensions = "10.15"
-jinja2 = "3.1.6"
 mike = "2.1.3"
 isort = "6.0.1"
 coveralls = "4.0.1"
@@ -51,6 +42,13 @@ ruff = "0.11.12"
 pytest-mock = "3.14.1"
 pre-commit = "4.2.0"
 mypy = "^1.15.0"
+
+[tool.poetry.scripts]
+experiment = "pysatl_experiment.cli.cli.cli:cli"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
 
 [tool.isort]
 line_length = 120
@@ -70,8 +68,10 @@ extend-ignore = [
   "E241", "E272", "E221", "B007", "B904", "S603", "S607", "S608", "NPY002"
 ]
 
-[tool.mypy]
-ignore_missing_imports = true
+[tool.ruff.lint.per-file-ignores]
+"tests/**/*.py" = [
+  "S101", "S104", "S311", "S105", "S106", "S110"
+]
 
 [tool.ruff.lint.mccabe]
 max-complexity = 12
@@ -79,16 +79,5 @@ max-complexity = 12
 [tool.ruff.lint.isort]
 lines-after-imports = 2
 
-[tool.flake8]
-ignore = ["E121", "E123", "E126", "E24", "E203", "E704", "W503", "W504"]
-max-line-length = 120
-max-complexity = 12
-exclude = [".git", "__pycache__", ".eggs", "user_data", ".venv", ".env"]
-
-[tool.ruff.lint.per-file-ignores]
-"tests/**/*.py" = [
-  "S101", "S104", "S311", "S105", "S106", "S110"
-]
-
-[tool.poetry.scripts]
-experiment = "pysatl_experiment.cli.cli.cli:cli"
+[tool.mypy]
+ignore_missing_imports = true
