+++ pysatl_experiment/__main__.py
@@ -1,0 +1,5 @@
+from pysatl_experiment.cli.cli.cli import cli
+
+
+if __name__ == "__main__":
+    cli()
